chore: move date locale store to lib
