+++ tests/unit/DateUtilsTest.ts
@@ -2,9 +2,9 @@
 import {addDays, addMinutes, format, setHours, setMinutes, subDays, subHours, subMinutes, subSeconds} from 'date-fns';
 import {toZonedTime, format as tzFormat} from 'date-fns-tz';
 import Onyx from 'react-native-onyx';
+import DateLocaleStore from '@libs/DateLocaleStore';
 import DateUtils from '@libs/DateUtils';
 import CONST from '@src/CONST';
-import DateLocaleStore from '@src/languages/DateLocaleStore';
 import TranslationStore from '@src/languages/TranslationStore';
 import ONYXKEYS from '@src/ONYXKEYS';
 import type {SelectedTimezone} from '@src/types/onyx/PersonalDetails';
